Lab1 Task 2 -->fix(1)
