+++ PetCare.Common/Entitys/BarkHandler.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace PetCare.Common.Entitys
+{
+    //Делегат для подій
+    public delegate void BarkHandler(string message);
+}
